requirements.txt
aiohttp==3.3.2
speex
pyyaml